# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Jobberwocky",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:251"
        }
    ],
    "paths": {
...
        "/api/skills": {
            "get": {
                "tags": [
                    "Skill"
                ],
                "summary": "Show all skills for jobs",
                "operationId": "7ef1b9241bba8e292fcb019dbe334fc0",
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/showSkill/{id}": {
            "get": {
                "tags": [
                    "Skill"
                ],
                "summary": "Displays the details of the skill",
                "operationId": "5c6d456c2c0865dd7fbb58d0352f442d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "skill identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/storeSkill": {
            "post": {
                "tags": [
                    "Skill"
                ],
                "summary": "Add a new Skill",
                "operationId": "0a0024c277d2dc36b11676c0adfbff20",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of the new skill",
                                        "format": "string",
                                        "default": "Laravel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/updateSkill": {
            "put": {
                "tags": [
                    "Skill"
                ],
                "summary": "Upgrade an existing skill.",
                "operationId": "ceb700eb78263885c7b8dd7f28ad6934",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of the skill",
                                        "format": "number",
                                        "default": "1"
                                    },
                                    "name": {
                                        "description": "name of the new skill",
                                        "format": "string",
                                        "default": "Laravel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/delSkill/{id}": {
            "delete": {
                "tags": [
                    "Skill"
                ],
                "summary": "Delete a selected skill",
                "operationId": "d60134cd57ce8528bb4ca8f9a06b5ad2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "skill identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}